pyproject.toml
[project]
name = "ytmusicstatus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "gunicorn",
    "python-dotenv==1.0.1",
    "uvicorn[standard]==0.34.0",
    "ytmusicapi==1.9.0"
]
